+++ Roll.py
@@ -1,15 +1,15 @@
 import asyncio
 
-from graia.application.exceptions import AccountMuted
 from graia.broadcast import Broadcast
-from graia.application.interrupt import InterruptControl
-from graia.application import GraiaMiraiApplication as Slave, Session, GroupMessage
+from graia.application.exceptions import AccountMuted
+from graia.application import Session, GraiaMiraiApplication as Slave
 
-from application.YummyPicture.Entrance import YummyPicture
-from application.TalkToMe.TalkToMe import TalkToMe
-from application.URaNai.URaNai import URaNai
-from application.Economy.EconomyCommand import Economy
-from application.Capitalism.Capitalism import Capitalism
+from application.YummyPicture.PictureRipperListener import PictureRipperListener
+from application.TalkToMe.TalkToMeListener import TalkToMeListener
+from application.URaNai.URaNaiListener import URaNaiListener
+from application.Economy.EconomyListener import EconomyListener
+from application.Capitalism.CapitalismListener import CapitalismListener
+from application.VideoRipper.VideoRipper import VideoRipperListener
 from Logger import logger
 
 loop = asyncio.get_event_loop()
@@ -17,23 +17,16 @@ loop = asyncio.get_event_loop()
 bcc = Broadcast(loop=loop)
 app = Slave(
     broadcast=bcc,
-    connect_info=Session(host="http://localhost:7737", authKey="", account=111111111, websocket=True)
+    connect_info=Session(host="http://localhost:7737", authKey="xxxx", account=1234, websocket=True)
 )
-inc = InterruptControl(bcc)
-
-
-@bcc.receiver("GroupMessage")
-async def group_message_handler(message: GroupMessage):
-    commands: [str] = message.messageChain.asDisplay().split(' ')
-    commands[0] = commands[0].upper()
-    try:
-        await YummyPicture(app, message, commands)
-        await TalkToMe(app, message, commands)
-        await URaNai(app, message, commands)
-        await Economy(app, message, commands)
-        await Capitalism(app, message, commands)
-    except AccountMuted:
-        logger.info(f'我被狗禁言了！！！')
-
+try:
+    TalkToMeListener(bcc).start()
+    URaNaiListener(bcc).start()
+    CapitalismListener(bcc).start()
+    EconomyListener(bcc).start()
+    VideoRipperListener(bcc).start()
+    PictureRipperListener(bcc).start()
+except AccountMuted:
+    logger.info(f'我被狗禁言了！！！')
 
 app.launch_blocking()
