+++ utils/network.py
@@ -33,9 +33,9 @@ async def requestText(url: str, method: str = 'GET', headers: dict = None, param
 
 
 async def request(method: str = 'GET', url: str = '', headers: dict = None, params: dict = None, body=None,
-                  connector: ProxyConnector = None, data=None, close=True):
+                  connector: ProxyConnector = None, data=None, cookies=None, close=True):
     async with aiohttp.request(method, url, headers=headers, params=params, json=body, connector=connector,
-                               data=data) as response:
+                               data=data, cookies=cookies) as response:
         raw_data = await response.read()
         if connector and close:
             await connector.close()
